Developing the UsersController class, creating the 'create,' 'show,' and 'delete' methods
Continuing the development of the application, I'm developing the UsersController class to manage actions related to user routes. Creating the 'create' method to create a user and record this user in the database, receiving the data of name, email, and password through the request. Finally, creating the 'show' and 'delete' methods to display information and delete a specific user, respectively, using the user's 'id' field passed in the request, in addition to creating these routes in the 'users.routes.js' file.