Developing the 'NotesController' class, creating the 'create,' 'show,' and 'delete' methods.
Continuing the development of the application, we are building the 'NotesController' class in the 'NotesController.js' file. We are creating the 'create' methods for adding a new note to the application, the 'show' method for displaying specific note information, and the 'delete' method for removing a particular note. Finally, we are utilizing the 'user.routes.js' file to establish routes related to notes within the application