Creating the functionality to list the registered movies and creating the tags route.
Adding the 'index' method to the 'NotesController' class, implementing the functionality to list and display registered movies, allowing filtering the search by the user's ID who wrote the note, the movie title, and the tags present in the movie. Also creating the tags route and the 'TagsController' class containing the logic to list registered tags filtered by the user's ID, through the 'index' method.