Creating the folder 'configs' and the configuration files 'uploadConfig.js' and 'authConfig.js', creating the folders 'providers' and 'middlewares' and the folders 'temp' and 'uploads'
Starting the final stretch of the application, evolving the project structure. Creating the 'configs' folder, responsible for storing general configuration files for the projects. Creating the 'uploadConfig.js' and 'authConfig.js' files inside the config folder, which will store configuration data related to image uploads and authentication, respectively. Creating the 'tmp' folder and its subfolder 'uploads', which will be used in the future to store image files that will be uploaded by users to serve as avatar images. Finally, creating the 'providers' and 'middlewares' folders and installing new dependencies in the project, the 'multer' library to handle file uploads, the 'jsonwebtoken' library for authentication, and the 'cors' library to work with Front-End integration.