Developing all the code responsible for the user authentication logic and preparing the environment for integration with the Front-End
Developing the code responsible for user validation logic entirely. Creating the 'SessionsController' to handle user sessions in the application, receiving the necessary user login data, email and password, through the request. If the email and password are correct, the response includes the data of the registered user and the authentication token used in subsequent requests. If the email or password is incorrect, an error is thrown. Creating the 'validateAuthentication' middleware in the 'middlewares' folder, which checks if the sent request has a header containing the authentication token. If the token is not present, an error is thrown indicating that the authentication token is required. If the token is present, the request is allowed to continue its route, passing the user id that made the request. Updating the 'notes.routes.js', 'tags.routes.js', and 'users.routes.js' files to implement the authentication middleware 'validateAuthentication' in the application routes. Creating the 'sessions.routes.js' file to manage the session route, responsible for user login in the application. Updating the 'TagsController', 'NotesController', and 'UsersController' controllers to receive the user id as data directly from the request, no longer as a request parameter. Finally, updating the database tables, testing the changes, and preparing the application for image upload logic