Developing the code responsible for the logic to upload the user's avatar image and save this data in the database
Continuing the final stage of application development. Developing the code responsible for the logic of updating the user's avatar image. Creating, within the 'providers' folder, the 'DiskStorage' class with the 'saveFile' and 'deleteFile' methods to save the user's image file in the designated folder and delete the old image during the update, respectively. Creating the 'AvatarController' with the responsibility of receiving the request and applying the logic to save the new user image, storing the image in the 'uploads' folder, and saving the path to the image in the database. Finally, creating the 'avatar.routes.js' route file, using the multer library to handle files, and modifying the 'server.js' file to create the '/files' route to display static files such as images