Preparing the application for deployment by installing the 'dotenv' and 'pm2' libraries
Preparing the application for deployment. Installing the 'dotenv' library to handle and store sensitive data through environment variables. Creating the '.env' file at the root of the project to store the environment variables that will be used in the application. Updating the 'server.js' and 'authConfig.js' files to use data from the environment variables. Updating the '.gitignore' file to ignore the '.env' file and creating the '.env.example' file to serve as an example file, containing the names of the variables being used but without the values. Installing the 'pm2' library to ensure that the application runs continuously in a production environment and is restarted in case of an error. Finally, creating the 'ecosystem.config.cjs' file, storing the necessary configurations for 'pm2,' and updating the 'package.json' file by changing the 'start' script to run the application using 'pm2'